chore(wasm): create the compact package generation
