+++ .github/workflows/make_release_tfhe.yml
@@ -20,6 +20,14 @@ on:
         description: "Push web compat (cross-origin) js package"
         type: boolean
         default: true
+      push_web_client_package:
+        description: "Push web client js package"
+        type: boolean
+        default: true
+      push_web_compat_client_package:
+        description: "Push web compat (cross-origin) client js package"
+        type: boolean
+        default: true
       push_node_package:
         description: "Push node js package"
         type: boolean
@@ -117,6 +125,40 @@ jobs:
 
       - name: Publish web compat (cross-origin) package
         if: ${{ inputs.push_web_compat_package }}
+        uses: JS-DevTools/npm-publish@0fd2f4369c5d6bcfcde6091a7c527d810b9b5c3f
+        with:
+          package: tfhe/pkg/package.json
+          dry-run: ${{ inputs.dry_run }}
+          tag: ${{ env.NPM_TAG }}
+          provenance: true
+
+      - name: Build web client package
+        if: ${{ inputs.push_web_client_package }}
+        run: |
+          rm -rf tfhe/pkg
+
+          make build_web_js_api_parallel_client
+          sed -i 's/"tfhe"/"tfhe-client"/g' tfhe/pkg/package.json
+
+      - name: Publish web client package
+        if: ${{ inputs.push_web_client_package }}
+        uses: JS-DevTools/npm-publish@0fd2f4369c5d6bcfcde6091a7c527d810b9b5c3f
+        with:
+          package: tfhe/pkg/package.json
+          dry-run: ${{ inputs.dry_run }}
+          tag: ${{ env.NPM_TAG }}
+          provenance: true
+
+      - name: Build web compat (cross-origin) client package
+        if: ${{ inputs.push_web_compat_client_package }}
+        run: |
+          rm -rf tfhe/pkg
+
+          make build_web_js_api_client
+          sed -i 's/"tfhe"/"tfhe-compat-client"/g' tfhe/pkg/package.json
+
+      - name: Publish web compat (cross-origin) client package
+        if: ${{ inputs.push_web_compat_client_package }}
         uses: JS-DevTools/npm-publish@0fd2f4369c5d6bcfcde6091a7c527d810b9b5c3f
         with:
           package: tfhe/pkg/package.json
