+++ tfhe/src/js_on_wasm_api/client/mod.rs
@@ -1,0 +1,12 @@
+mod config;
+mod integers;
+mod keys;
+#[cfg(feature = "zk-pok")]
+mod zk;
+
+pub(crate) use config::*;
+#[cfg(feature = "integer-client-js-wasm-api")]
+pub(crate) use integers::*;
+pub(crate) use keys::*;
+#[cfg(feature = "zk-pok")]
+pub(crate) use zk::*;
