+++ tfhe/src/js_on_wasm_api/js_high_level_api/keys.rs
@@ -1,13 +1,8 @@
 use crate::high_level_api as hlapi;
-use crate::js_on_wasm_api::js_high_level_api::config::TfheConfig;
-use crate::js_on_wasm_api::js_high_level_api::{catch_panic, catch_panic_result, into_js_error};
+use crate::js_on_wasm_api::client::{TfheClientKey, TfheCompactPublicKey};
 use crate::js_on_wasm_api::shortint::ShortintCompactPublicKeyEncryptionParameters;
+use crate::js_on_wasm_api::{catch_panic, catch_panic_result, into_js_error};
 use wasm_bindgen::prelude::*;
-
-#[wasm_bindgen]
-pub fn init_panic_hook() {
-    console_error_panic_hook::set_once();
-}
 
 /// Initialize a cross-origin worker pool for parallel computation.
 ///
@@ -67,56 +62,6 @@ pub async fn init_cross_origin_worker_pool_from_worker(
     wasm_par_mq::init_pool_sync_from_worker(num_workers)
         .await
         .map_err(|e| JsValue::from_str(&e))
-}
-
-#[wasm_bindgen]
-pub struct TfheClientKey(pub(crate) hlapi::ClientKey);
-
-#[wasm_bindgen]
-impl TfheClientKey {
-    #[wasm_bindgen]
-    pub fn generate(config: &TfheConfig) -> Result<TfheClientKey, JsError> {
-        catch_panic(|| Self(hlapi::ClientKey::generate(config.0)))
-    }
-
-    #[wasm_bindgen]
-    pub fn generate_with_seed(
-        config: &TfheConfig,
-        seed: JsValue,
-    ) -> Result<TfheClientKey, JsError> {
-        catch_panic_result(|| {
-            let seed =
-                u128::try_from(seed).map_err(|_| JsError::new("Value does not fit in a u128"))?;
-            let key = hlapi::ClientKey::generate_with_seed(config.0, crate::Seed(seed));
-            Ok(Self(key))
-        })
-    }
-
-    #[wasm_bindgen]
-    pub fn safe_serialize(&self, serialized_size_limit: u64) -> Result<Vec<u8>, JsError> {
-        let mut buffer = vec![];
-        catch_panic_result(|| {
-            crate::safe_serialization::SerializationConfig::new(serialized_size_limit)
-                .serialize_into(&self.0, &mut buffer)
-                .map_err(into_js_error)
-        })?;
-
-        Ok(buffer)
-    }
-
-    #[wasm_bindgen]
-    pub fn safe_deserialize(
-        buffer: &[u8],
-        serialized_size_limit: u64,
-    ) -> Result<TfheClientKey, JsError> {
-        catch_panic_result(|| {
-            crate::safe_serialization::DeserializationConfig::new(serialized_size_limit)
-                .disable_conformance()
-                .deserialize_from(buffer)
-                .map(Self)
-                .map_err(into_js_error)
-        })
-    }
 }
 
 // Wasm cannot generate a normal server key, only a compressed one
@@ -258,57 +203,6 @@ impl TfheCompressedPublicKey {
             crate::safe_serialization::DeserializationConfig::new(serialized_size_limit)
                 .disable_conformance()
                 .deserialize_from(buffer)
-                .map(Self)
-                .map_err(into_js_error)
-        })
-    }
-}
-
-#[wasm_bindgen]
-pub struct TfheCompactPublicKey(pub(crate) hlapi::CompactPublicKey);
-
-#[wasm_bindgen]
-impl TfheCompactPublicKey {
-    #[wasm_bindgen]
-    pub fn new(client_key: &TfheClientKey) -> Result<TfheCompactPublicKey, JsError> {
-        catch_panic(|| Self(hlapi::CompactPublicKey::new(&client_key.0)))
-    }
-
-    #[wasm_bindgen]
-    pub fn safe_serialize(&self, serialized_size_limit: u64) -> Result<Vec<u8>, JsError> {
-        let mut buffer = vec![];
-        catch_panic_result(|| {
-            crate::safe_serialization::SerializationConfig::new(serialized_size_limit)
-                .serialize_into(&self.0, &mut buffer)
-                .map_err(into_js_error)
-        })?;
-
-        Ok(buffer)
-    }
-
-    #[wasm_bindgen]
-    pub fn safe_deserialize(
-        buffer: &[u8],
-        serialized_size_limit: u64,
-    ) -> Result<TfheCompactPublicKey, JsError> {
-        catch_panic_result(|| {
-            crate::safe_serialization::DeserializationConfig::new(serialized_size_limit)
-                .disable_conformance()
-                .deserialize_from(buffer)
-                .map(Self)
-                .map_err(into_js_error)
-        })
-    }
-
-    #[wasm_bindgen]
-    pub fn safe_deserialize_conformant(
-        buffer: &[u8],
-        serialized_size_limit: u64,
-        conformance_params: &ShortintCompactPublicKeyEncryptionParameters,
-    ) -> Result<TfheCompactPublicKey, JsError> {
-        catch_panic_result(|| {
-            crate::safe_serialization::DeserializationConfig::new(serialized_size_limit)
-                .deserialize_from(buffer, &conformance_params.compact_pke_params)
                 .map(Self)
                 .map_err(into_js_error)
         })
