+++ tfhe/src/js_on_wasm_api/mod.rs
@@ -1,3 +1,5 @@
+pub(crate) mod client;
+
 #[cfg(feature = "shortint-client-js-wasm-api")]
 mod shortint;
 
@@ -15,4 +17,22 @@ mod js_high_level_api;
 
 pub(crate) fn into_js_error<E: std::fmt::Debug>(e: E) -> wasm_bindgen::JsError {
     wasm_bindgen::JsError::new(format!("{e:?}").as_str())
+}
+
+pub(crate) fn catch_panic_result<F, R>(closure: F) -> Result<R, wasm_bindgen::JsError>
+where
+    F: FnOnce() -> Result<R, wasm_bindgen::JsError>,
+{
+    std::panic::catch_unwind(std::panic::AssertUnwindSafe(closure))
+        .unwrap_or_else(|_| Err(wasm_bindgen::JsError::new("Operation Failed")))
+}
+
+pub(crate) fn catch_panic<F, R>(closure: F) -> Result<R, wasm_bindgen::JsError>
+where
+    F: FnOnce() -> R,
+{
+    std::panic::catch_unwind(std::panic::AssertUnwindSafe(closure)).map_or_else(
+        |_| Err(wasm_bindgen::JsError::new("Operation Failed")),
+        |ret| Ok(ret),
+    )
 }
